Move skinned mesh debug to system
